move `LoadedDocumentContent` interface to types.ts
