+++ src/commands/topics.ts
@@ -27,7 +27,8 @@ import { isCCloud } from "../models/resource";
 import { Schema } from "../models/schema";
 import { KafkaTopic } from "../models/topic";
 import { schemaKindMultiSelect, SchemaKindSelection } from "../quickpicks/schemas";
-import { loadDocumentContent, LoadedDocumentContent, uriQuickpick } from "../quickpicks/uris";
+import { LoadedDocumentContent } from "../quickpicks/types";
+import { loadDocumentContent, uriQuickpick } from "../quickpicks/uris";
 import { promptForSchema } from "../quickpicks/utils/schemas";
 import { getSubjectNameStrategy } from "../quickpicks/utils/schemaSubjects";
 import { JSON_DIAGNOSTIC_COLLECTION } from "../schemas/diagnosticCollection";
