+++ src/quickpicks/types.ts
@@ -1,4 +1,4 @@
-import { QuickPickItem } from "vscode";
+import { QuickPickItem, TextDocument } from "vscode";
 
 /**
  * Extension of {@link QuickPickItem} to include an associated `value` for easier matching on
@@ -8,4 +8,11 @@ import { QuickPickItem } from "vscode";
 export type QuickPickItemWithValue<T> = QuickPickItem & {
   value?: T | undefined;
 };
-// TODO(shoup): migrate other quickpicks to use this type
+
+/** Representation of content retrieved from a file or editor. `openDocument` will be provided if
+ * the content came from an open editor, or if the associated file is open in an editor for the
+ * current workspace. */
+export interface LoadedDocumentContent {
+  content: string;
+  openDocument?: TextDocument;
+}
